refactoring and extraction of departmentHelper.js
