+++ src/modules/apiHelpers/fikaHelper.js
@@ -11,14 +11,4 @@ const Fika = {
   },
 };
 
-const Department = {
-  async index() {
-    let response = await axios.get("/api/departments");
-    store.dispatch({
-      type: "SET_DEPARTMENT_INDEX",
-      payload: response.data.departments,
-    });
-  },
-};
-
-export { Fika, Department };
+export { Fika };
